Commented out non-working clearConsole method, need suggestions!
Hi GitHub community! I'm a noob programmer looking back to a school project I submitted a while ago, one which is functional but left some things to be desired. 

I figured since I'm new to the community I could both look to improve this program here for not only the sake of improving my craft but also for understanding the GitHub flow process and tools. Two birds, one stone as they say!

Any suggestions would be greatly appreciated! Have enjoyed the community immensely so far and I hope to dive deeper in the near future....cheers!